hugme/requirements.txt
discord.py
httpx
SQLAlchemy
asyncpg
python-dotenv
aiohttp